pyproject.toml
[project.urls]
Repository = "https://github.com/nachatz/pytest-streaming"

[project]
name = "pytest-streaming"
version = "0.1.0"
description = "Plugin for testing pubsub, pulsar, and kafka systems with pytest locally and in ci/cd"
readme = "README.md"
authors = [{ name = "Nikolas Achatz", email = "nil@gmail.com" }]
maintainers = [{ name = "Nikolas Achatz", email = "nil@gmail.com" }]
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Pytest",
]
dependencies = [
    "google-cloud-pubsub>=2.29.0",
    "httpx>=0.28.1",
    "pulsar-client>=3.5.0",
    "pytest>=8.3.5",
    "types-protobuf>=5.29.1.20250315",
]

[tool.setuptools.packages.find]
where = ["pytest_streaming"]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.9",
    "mypy>=1.15.0",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.11.2",
    "coverage>=7.7.1",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.3",
    "dotenv>=0.9.9",
    "pytest-vcr>=1.0.2",
]

[project.entry-points.pytest11]
pytest_streaming = "pytest_streaming.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
source = ["pytest_streaming"]
omit = ["config.py"]

[tool.coverage.report]
fail_under = 100
show_missing = true


